requirements.txt
astroid==2.5.2
bidict==0.21.2
cffi==1.14.5
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-SocketIO==5.0.1
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.20
pylint==2.7.4
python-dotenv==0.17.0
python-engineio==4.0.1
python-socketio==5.1.0
toml==0.10.2
typed-ast==1.4.2
Werkzeug==1.0.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.3.0
